Implement tab page stack save/restore
This change implements tab page stack state
saving/restoration. Essentially, page types are augmented with an
interface method that, when called, returns a serialized function that
is capable of reconstructing the page based on its current state. It
is the responsibility of page implementations to ensure their own
correct reconstruction. When the fragment managing tabs is destroyed,
it asks all tabs to serialize their page stacks. When the fragment
managing tabs is restored, it asks all tabs to restore their page
stacks using the previously serialized state.

This all appears to work correctly, but it does highlight a weakness
in the current implementation of page creation: Page creation should
be generalized so that a page can say:

  * "Open the given page on the same tab as me"
  * "Open the given page on a specific tab"

In both cases, if the tab that will contain the new page is not
currently selected, the tab view should appear to navigate to that tab
to show the page. In the current code, page creation is essentially
fixed: Specific pages can only be opened on specific tabs, and if
the tab in question isn't the currently selected tab, the page still
appears but the tab view isn't updated and so it appears that the
page is on the wrong tab (clicking any of the tabs "fixes" the issue).